# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/active-channels": {
      "post": {
        "operationId": "ActiveChannelsController_create",
        "summary": "Create a new active channel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActiveChannelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The channel has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveChannel"
                }
              }
            }
          }
        },
        "tags": [
          "active-channels"
        ]
      },
      "get": {
        "operationId": "ActiveChannelsController_findAll",
        "summary": "Get all active channels",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all active channels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveChannel"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "active-channels"
        ]
      }
    },
    "/active-channels/search": {
      "get": {
        "operationId": "ActiveChannelsController_search",
        "summary": "Search channels by filters",
        "parameters": [
          {
            "name": "megagroup",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "reactRestricted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "banned",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "reactions",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "availableMsgs",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "dMRestriction",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "wordRestriction",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "username",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sendMessages",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "restricted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "participantsCount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "canSendMsgs",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "broadcast",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "channelId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return matching channels.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveChannel"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "active-channels"
        ]
      }
    },
    "/active-channels/{channelId}": {
      "get": {
        "operationId": "ActiveChannelsController_findOne",
        "summary": "Get an active channel by channelId",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the active channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveChannel"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found"
          }
        },
        "tags": [
          "active-channels"
        ]
      },
      "patch": {
        "operationId": "ActiveChannelsController_update",
        "summary": "Update an active channel by channelId",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActiveChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The channel has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveChannel"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found"
          }
        },
        "tags": [
          "active-channels"
        ]
      },
      "delete": {
        "operationId": "ActiveChannelsController_remove",
        "summary": "Delete an active channel by channelId",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The channel has been successfully deleted."
          },
          "404": {
            "description": "Channel not found"
          }
        },
        "tags": [
          "active-channels"
        ]
      }
    },
    "/active-channels/{channelId}/reactions": {
      "post": {
        "operationId": "ActiveChannelsController_addReaction",
        "summary": "Add reaction to chat group",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "active-channels"
        ]
      },
      "delete": {
        "operationId": "ActiveChannelsController_removeReaction",
        "summary": "Remove reaction from chat group",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "active-channels"
        ]
      }
    },
    "/active-channels/{channelId}/reactions/random": {
      "get": {
        "operationId": "ActiveChannelsController_getRandomReaction",
        "summary": "Get a random reaction from chat group",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "active-channels"
        ]
      }
    }
  },
  "info": {
    "title": "NestJS and Express API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateActiveChannelDto": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string"
          },
          "broadcast": {
            "type": "boolean",
            "default": false
          },
          "canSendMsgs": {
            "type": "boolean",
            "default": true
          },
          "participantsCount": {
            "type": "number",
            "default": 300
          },
          "restricted": {
            "type": "boolean",
            "default": false
          },
          "sendMessages": {
            "type": "boolean",
            "default": true
          },
          "reactRestricted": {
            "type": "boolean",
            "default": false
          },
          "title": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "wordRestriction": {
            "type": "number",
            "default": 0
          },
          "dMRestriction": {
            "type": "number",
            "default": 0
          },
          "availableMsgs": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reactions": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "banned": {
            "type": "boolean",
            "default": false
          },
          "megagroup": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "channelId",
          "broadcast",
          "canSendMsgs",
          "participantsCount",
          "restricted",
          "sendMessages",
          "reactRestricted",
          "title",
          "username",
          "wordRestriction",
          "dMRestriction",
          "availableMsgs",
          "reactions",
          "banned"
        ]
      },
      "ActiveChannel": {
        "type": "object",
        "properties": {}
      },
      "UpdateActiveChannelDto": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string"
          },
          "broadcast": {
            "type": "boolean",
            "default": false
          },
          "canSendMsgs": {
            "type": "boolean",
            "default": true
          },
          "participantsCount": {
            "type": "number",
            "default": 300
          },
          "restricted": {
            "type": "boolean",
            "default": false
          },
          "sendMessages": {
            "type": "boolean",
            "default": true
          },
          "reactRestricted": {
            "type": "boolean",
            "default": false
          },
          "title": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "wordRestriction": {
            "type": "number",
            "default": 0
          },
          "dMRestriction": {
            "type": "number",
            "default": 0
          },
          "availableMsgs": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reactions": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "banned": {
            "type": "boolean",
            "default": false
          },
          "megagroup": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "AddReactionDto": {
        "type": "object",
        "properties": {
          "reactions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "reactions"
        ]
      }
    }
  }
}